+++ sample/src/main/kotlin/ca/allanwang/kau/sample/KPrefSample.kt
@@ -8,6 +8,7 @@ import ca.allanwang.kau.kpref.kpref
  * Created by Allan Wang on 2017-06-07.
  */
 object KPrefSample : KPref() {
+    var version: Int  by kpref("version", -1)
     var textColor: Int by kpref("TEXT_COLOR", Color.WHITE)
     var accentColor: Int by kpref("ACCENT_COLOR", 0xffff8900.toInt())
     var bgColor: Int by kpref("BG_COLOR", 0xff303030.toInt())